refactor(types): move Threshold + ThresholdStates to src/types.ts (refs #298)
