+++ src/components/options/thresholds/types.ts
@@ -1,19 +1,6 @@
-import { SelectableValue } from '@grafana/data';
-
-export interface Threshold {
-  color: string;
-  state: number;
-  value: number;
-}
+import { Threshold } from 'types';
 
 export interface ThresholdItemTracker {
   threshold: Threshold;
   ID: string;
 }
-
-export const ThresholdStates: SelectableValue[] = [
-  { value: 0, label: 'ok' },
-  { value: 1, label: 'warning' },
-  { value: 2, label: 'critical' },
-  { value: 3, label: 'custom' },
-];
